requirements.txt
transformers
datasets
sentencepiece
pytorch-lightning
Pillow
protobuf == 3.20
nltk
Flask
flask-cors
